refactor orderbook
